feat: add static signals to handle list changes
Signals for list changes are called by the GUI. This means I don't have
to constantly connect/destroy the models signals when the list/item
changes.

The signals are static and setup by the constructor so that all GUI
objects in the application can catch them.

The signal includes the item being edited so the GUI can test to see if
an update is necessary.