feat: Start work on the menubar
This was a lot more complicated than I expected it to be.

I had to create an ApplicationWindow and Application to register the
Actions (I tried 'simpler' tutorials but they didn't work either)

The Editor and Window are separate because I think their classes are
going to be huge in the future.