refactor: get rid of shared_ptr in models
I found out I was an idiot and didn't need a shared_ptr<> to implement
the undo stack. A simple unique_ptr<> would have worked fine.

This commit removes all of the shared_ptr<>/weak_ptr<> in the models and
replaces them with references and unique_ptr<>s

The GUI code was changed so that compiles. Signals and variables that
were shared_ptr<> are now a regular pointer. To ensure that there is no
access after delete the selection is cleared before deleting an item.