feat: process SI overlapping frame objects
This mess of code (I tried to make it as readable as possible) will
detect and process two overlapping sprites in the sprite importer file.

This use case for frames that are similar to each other but have,
either:

 * a small change in a small tile overlapping a large one, or
 * a small change that overlaps multiple tiles (rocket jump frame -
   adding fire to the armour exhaust port)

Testing involved a simple sprite sheet with 32 frames arranged in a 2x3
grid of 8x8 tiles. Two of the sprites (blink & wink) were very similar
to stand except for pixels in the face area that covered 4 sprites (see
figure). After marking the changes in utsi file, the number of tiles
produced by `utsi2utms` changed from 113 to 104.

                              --- ---
                             |   |   |
                             |  x|x  |
                              --- ---
                             |  x|x  |
                             |   |   |
                              --- ---
                             |   |   |
                             |   |   |
                              --- ---
                        Placement of overtile