fix: Prevent intermittent segfault on frame Destroy
The segfault was caused by wxNotebook emitting a
wxEVT_NOTEBOOK_PAGE_CHANGING event when it is being Destroyed and the
user is not on the first tab.

I have moved the `_controller.emitAllDataChanged` call to a timer event
that is executed 100ms after wxFrame is shown and rendered to the
display.

This is a dirty hack, but it works until I can find a better fix. Users
should not notice this because the widgets that need to be disabled by
UpdateGui are on the second/third page of the wxNotebook.