pod refactor: MetaSprite / SpriteImporter mode data
Replaces the MetaSprite/SpriteImporter model data with POD structs.

This is a LOT neater and uses less code.
Validation will be done with a isValid() function call in the
GUI/Processor. It will just be less hassle on my part.

Also moved SpriteImporter into MetaSprite folder to reduce namespace
pollution.

This code is missing the export order document, I still need to think
about how to tackle that.