feat: Merge Undo Actions
The Container Controllers can now retrieve the previous undo action if
it matches the current ActionType, is merge-able and the undoStack
allows it. The controller will then reuse the previous Action instance
to store Undo State.

In order to prevent mouse actions from being merged in the Graphical
editors `undoStack.dontMergeNextAction` is called on OnMouseLeftDown.

In order to prevent the system from merging separate text entries that
share an action (ie, aabb, points, grid, location), a new class,
DontMergeFocusHack, was created to recursively search through all the
widgets and setup an event that calls `undostack.dontMergeNextAction`
whenever a wxTextCtrl or a wxSpinCtrl is focused.