perf: Replace SI FrameSet image with a shared_ptr
This causes the SI FrameSet copy constructor to copy a pointer instead
of the entire image, saving ~50-500K of memory (depending on image size)
per FrameSet edit in the undo stack.

This also allows SI::FrameSet::reloadImage to affect ALL images in the
undo stack instead of just the selected FrameSet.