fix: propagate tab and focus-out events in aabb widgets
Previously the FrameContentsDelegate would not emit the commitData or
closeEditor signals when focusing out of a PointWidget or RectWidget.

In order to emit these signals QAbstractItemDelegate requires the widget
created by the createEditor virtual function to receive tab-press and
focus-out events. Unfortunately these events are sent to the QSpinBox
inside the PointWidget/RectWidget and therefore are not caught by
QAbstractItemDelegate.

This commit catches the spinbox tab/focus-out events and resends them to
to the aabb widget so they can be caught by QAbstractItemDelegate.