feat: untech-write-sfc-checksum command line utility
untech-write-sfc-checksum is a command line utility that calculates and
fixes the internal checksum of an unheadered .sfc ROM file.

In order to prevent untech-write-sfc-checksum from corrupting an
innocent file it will only accept an .sfc file with an internal header
that matches the one created by untech-engine's `snes_header.inc`.