Fix segfault when adding a filename to a PropertyModel
This segfault occurs on my Windows 7 VM on Qt 5.8 with g++ 5.3.0
i686-w64-mingw32.

I am unsure what is causing the segfault, my best guess is that Qt 5.8
or Windows is corrupting the stack. It only occurs in the
PropertyListView, the other QFileDialog usages do not segfault.

This segfault still occurs when compiling with no LTO and -O1. I've also
tried changing the dialog's parent to nullptr, but it still segfaults.

I've discovered that using the QFileDialog::DontUseNativeDialog option
bypasses the windows dialog and prevents the segfault from occurring.

For consistency QFileDialog::DontUseNativeDialog has been applied to
every QFileDialog in the editor.