Fix bug when a RI external file is deleted and then created
Previously there existed the possibility that when an external file is
deleted and then created (ie, git checkout -f) the ImageCache will not
load the newly created file, instead returning a blank image.

This commit fixes the bug by invalidating the ImageCache and marking the
ResourceItem dirty if the external file exists and is not being watched.
This check occurs whenever the selected resource item is changed or the
external file list is changed.

If the user still encounters this bug, then all they will need to change
the selected ResourceItem and then select the bugged-out item.