Improve CombineSmallTiles by testing blank tiles
The old behaviour did not test the blank tiles when determining which
tiles to pair-off.  This was causing the first pass to produce
non-optimal results.

This commit pads the small tile graph to a multiple of four with blank
tiles, ensuring the blank tiles are accounted for when pairing tiles.

As a consequence the new code is simpler as there is always an even
number of matches.