Optimize MtTileset painting by caching pixmap fragments
This commit optimizes MetaTile Tileset painting by caching the pixmap
fragments vector inside the `MtGridGraphicsItem` and only updating it
when the grid or the number of MetaTiles changes.

A new class, `MtTilesetGridPainter`, has been created for the fragment
cache and painting function so it can be easily reused in the future.