Fix MS shift tileIds when adding/removing a tile
Previously the system would only edit the tileIds of the current frame
when adding/removing a tile. Investigation also discovered that the
possible for the shift tile code to underflow the tileId.

This commit fixes the shift frame object code so it processes all frame
objects are includes an unsigned underflow check.