Refactor RES::ErrorList into a generic ErrorList
The UnTech::Resources::ErrorList class has been moved to
UnTech::ErrorList.

This new class is completely generic.  Errors and warnings can either be
a string or a subclass of AbstractSpecializedError.

This commit also refactors the `std::vector<InvalidImageTile>`
specialization from the old ErrorList into a InvalidImageError class.
However, as each AbstractSpecializedError is represented as a single
error in the GUI, the AnimatedTileset compiler has been modified to
create a new error for each invalid image/animation.