Remove bounds checks from from SI::FrameSetReader
After testing the previous commit I noticed that the FrameSetReader was
raising a "number too large" xml_error on files that were created by the
GUI and were saved successfully.

As this is unexpected behaviour, I have removed the bounds checks from
FrameSetReader.  This will allow the GUI to open utsi files with invalid
frames and allow the user to correct the bad input with the GUI.