Optimize Resource Item data access
This commit adds a reference to the resource item container to the
various Resource Item classes, reducing the indirection to retrieve an
item from 3 pointer dereferences to one.

This commit will also speed up the compile time as
`models/project/project.h` is no longer required when including a
resource item class.