Change MS data format to replace addresses with indexes
This commit changes the MetaSprite data format, replacing the majority
of the address fields with Index+1 or Word Index fields.

PaletteList and DMA_Tile16Data still use address fields as the system
requires

As the untech-editor now disallows null FrameSets the FrameSetList has
also been removed.

The new format is simpler and will allow the untech-compiler to output a
binary data block instead of a bass assembler stream.