Save compiled MS data to the binary file
This commit changes the MetaSprite compiler and ProjectCompiler to store
the majority of the CompiledRomData in the output binary file.

The `RomDmaTile16Entry`s and the paletteList are still written to the
output inc file, but are now saved in the `RES_Lists` rodata block under
the Project namespace.

All MetaSprite data has been moved to the Project namespace in the
output inc file.