Refactor NamedList accessors into an abstract class
This commit creates two abstract classes AbstractListAccessor and
AbstractNamedListAccessor which hold the common actions used by the
NamedList accessor classes.

This new design allows me to remove the stupid template lambda thing in
NamedListModel and NamedListView and replace it with a saner virtual
function call.

The use of a template class allows me the compiler to make the usual
create, update, modify, delete with minimal hassle.