Refactor selection management code into ListUndoHelper
This commit changes the ListUndoHelper so is now responsible for
managing the accessor class selection when:

 * An item is added
 * An item is removed, or
 * An item is moved

This used to be the responsibility of the accessor class but the
interactions of the various signals and slots would occasionally cause
other subsystems (usually QAbstractItemView) to mangle the selection.

By moving the selection change into the Undo Command (after the list
signals have been emitted) we ensure the same item remains selected
after the QUndoCommand is invoked.

I initially tried to use SFINAE to have the ListUndoHelper automatically
decide which SelectionModifier to use, but it was unable to detect the
existence of a `selectedIndex` or `selectedIndexes` method in the
template accessors and kept falling back to BlankSelectionModifier.
Thus this design uses the more reliable (but uglier) method of using two
template arguments for ListUndoHelper.

This commit also removes the SelectedIndexHelper and
MultipleSelectedIndexesHelper classes, as they are no longer need.