Mark RI dependents unchecked if name changed or state changed
This commit moves `AbstractResourceItem::markDependantsUnchecked` to
`ResourceValidationWorker::markItemDependantsUnchecked` and invokes it
when:

 * A resource item is loaded
 * A resource item is compiled
 * Just before a resource item's name is changed
 * After the resource item is validated
 * The resource item state is changed to an error state
 * A resource item is about to be removed

By invoking `markItemDependantsUnchecked` twice whenever the name is
changed (once before `AbstractResourceItem::_name` is set and again
afterwards), the system can invalidate any resource item that references
the old or new name as a dependency.

Moving the `markItemDependantsUnchecked` function to the
`ResourceValidationWorker` class would make it easier to create a
dependants lookup map in the future.