Change MtTileset scratchpad to a grid<uint8_t>
As the MetaTile Tileset is now 256 MetaTiles in size, the scratchpad/map
can now be stored in a byte grid.

This commit is in preparation of the MetaTile map editor, which will
also use a grid<uint8_t> to store the map data.

The MtGraphicsScene selection grid is still a grid<uint16_t> as we need
a way to differentiate between solid and transparent tiles.