User can now select tiles in MtEditableGraphicsScene
Previously the MtEditableGraphicsScene was unable to select tiles with
the mouse as the tile selection code interfered with the
TileCursorGraphicsItem mouse event code.

This commit modifies the MtGraphicsScene::gridSelectionEdited signal so
it is only emitted upon any of the following conditions:

 * The user clicks on a single tile without pressing Control
 * The user lets go of Control
 * The user moves the mouse outside the MtGraphicsScene, or
 * The MtGraphicsScene looses focus

With this new modification, there is no-longer any interference between
the TileCursorGraphicsItem and MtGridGraphicsItem mouse event handlers
and the user can now select tiles in the MtEditableGraphicsScene.