Change RomDataWriter so it is memory address aware
This commit removes the old `Project::BlockSettings` and replaces it
with a `Project::MemoryMapSettings` class.  MemoryMapSettings allows
the new RomDataWriter to determine the addresses of inserted data.

untech-compiler no longer uses a project definable block size and
instead uses a block size based on the selected mapping mode (64KiB for
HiROM, 32KiB for LoROM)