Fix unix File::atomicWrite not detecting symbolic links
On unix systems, `File::atomicWrite` is supposed to throw an exception
if the filename is a symbolic link.  Testing revealed that it was
erroneously ignoring symlinks and overwriting them (deleting the symlink
and replacing it with a regular file).

Researching the issue revealed the `stat` function follows symlinks.
Replacing it with `lstat` (which does not follow links) fixes the issue.