Different namespaces for player and projectile entity function tables
This commit adds an entityType field to the EntityFunctionTable struct
to declare if a function table is for a player, projectile or entity.

The compiler has been changed to write the entity function table data
and defines to the following namespaces:

 * Entities - for entities
 * Entities.Projectiles - for projectiles
 * Entities.Players - for players

The `Entities.Projectiles` and `Entities.Players` namespaces were chosen
to allow the projectiles and players easy access to the macros within
the `Entities` namespace.