Cleanup selection code
All calls to `UpdateSelection` now occur at the end of the
`processGui()` method.  This will prevent glitches or bad state when a
selection is used in two different windows.

`MultipleChildSelection` is no-longer a subclass of `MultipleSelection`.
This will cause a compiler error if I do not include the parent selector
in `UpdateSelection`.