Fix ProjectListWindow::hasPendingActions()
This commit fixes a bug where the current editor is repeatedly closed
and reopened whenever a ProjectListWindow popup is open.

`hasPendingActions()` should only return true if
`processPendingActions()` will actually do something.

This commit also moves the location of the `processPendingActions()`
call as it should not be called if `hasPendingActions()` returned
false.