Remove implicit std::string to idstring conversion
I noticed the room script compiler was implicitly converting
`std::string` to `idstring` in order to find the id number of the
various instruction arguments.  This involves an unnecessary call to
`idstring::idValid(const string&)` that could be eliminated if the
search functions and name maps used `std::string` instead.

Thinking about it a bit more, I decided the only time I want to convert
a `std::string` to an `idstring` is during variable initialisation and
in the serializer.

This commit makes the `idstring(std::string)` constructor explicit and
removes the `idstring::operator=(const std::string&)` method.

In order for the code to compile, this commit also changes the name
search functions and name maps to use `std::string` instead of
`idstring`.