Fix missing Texture move constructor
The old code worked because the compiler skipped the Texture destructor
with copy elision.  Disabling copy elision with
`-fno-elide-constructors` causes a use-after-free bug and missing
textures on the About UnTech Editor window.

This commit implements the move constructors in the Texture and Texture8
classes.

This commit has been tested with `-fno-elide-constructors`.