Refactor to clear valid flags using the undo subsystem
This commit moves the responsibility of clearing the various EditorGui
valid flags into the undo action classes.

Valid flags are declared by adding a constexpr static member pointer
named `validFlag` to an action class's ActionPolicy.  The flag will be
cleared after the action changes editor data.

This commit also removes the call to `AbstractEditorGui::resetState()`
after an undo or redo as it is no longer needed.