Rewrite Image and IndexedImage
This rewrite stores the image data inside a std::unique_ptr<T[]>.
Fixing a [cppcoreguidelines-owning-memory] clang-tidy warning in the
Image and IndexedImage class declarations.

Also improves the memory safety of the constructors by checking the
`size` argument before `_imageData` is allocated.

Also fixes a memory leak in `IndexedImage::loadPngImage_shared()` that
can occur if the png file can be decoded and the colortype is not
LCT_PALETTE.